+++ ext_tables.sql
@@ -1,3 +1,4 @@
 CREATE TABLE sys_domain (
-    domainName VARCHAR(255) NOT NULL DEFAULT ''
+    domainName                            VARCHAR(255) NOT NULL DEFAULT '',
+    tx_bartacusplatformsh_routeDomainName VARCHAR(255) NOT NULL DEFAULT ''
 );